requirements.txt
Django==1.8.3
dj-database-url==0.3.0
project-runpy==0.3.1
django-extensions==1.5.5

boto==2.38.0
dj-obj-update==0.2.1


# Deployment
whitenoise==2.0.2
waitress==0.8.9


# Testing
factory-boy==2.5.2
mock==1.3.0
